fix: 修复基线缺陷
1. GROUP BY 输出顺序不确定：groupByElicit 中对分组键排序后迭代，
   聚合函数列同步使用排序后的键列表，确保输出可重现。
2. COUNT(*) 等表达式在 groupByElicit 中因 getSelectColumn() 无法
   提取列节点导致 IndexOutOfBoundsException，增加空列表保护。
3. Formula.addToDataList 中 int/long/short 类型使用 Integer.parseInt
   等直接解析，聚合函数返回的 Double 值（如 "3000.0"）导致
   NumberFormatException，统一改为 Double.parseDouble 中转。
4. 相应更新 3 个受排序影响的回归测试期望值。
5. UI 输入解析重构（execute_UI）：整块粘贴的 SQL 按分号切分
   （引号感知，忽略单引号字符串内的分号），语句内允许换行；
   无法整体解析且含换行的段逐行累积，能完整解析即执行，兼容
   不带分号、每行一条命令的输入。
6. 注释处理：仅当整段所有非空行均为 -- 注释时才跳过；
   注释行后跟真实语句时交由解析器执行（语句内注释被忽略）。
7. 其他基础修复：新增 rst 快捷指令（等同 resetdb）；错误信息
   压缩为一行并去掉类名前缀；execute() 捕获 TokenMgrError
   （中文分号等词法错误不再导致崩溃）；drop class exist 改为
   best-effort 清理，避免每次执行报错。

+++ app/src/main/java/edu/whu/tmdb/query/operations/utils/Formula.java
@@ -205,13 +205,13 @@ public class Formula {
                 dataList.add(Float.parseFloat(String.valueOf(obj)));
                 break;
             case "int":
-                dataList.add(Integer.parseInt(String.valueOf(obj)));
+                dataList.add((int) Double.parseDouble(String.valueOf(obj)));
                 break;
             case "long":
-                dataList.add(Long.parseLong(String.valueOf(obj)));
+                dataList.add((long) Double.parseDouble(String.valueOf(obj)));
                 break;
             case "short":
-                dataList.add(Short.parseShort(String.valueOf(obj)));
+                dataList.add((short) Double.parseDouble(String.valueOf(obj)));
                 break;
             case "double":
                 dataList.add(Double.parseDouble(String.valueOf(obj)));
